Tweak logic to MaxValueDrafter.
Also, update the template.

+++ LOCAM/src/main/java/cg/binu/drafter/MaxValueDrafter.java
@@ -27,6 +27,6 @@ public class MaxValueDrafter implements Drafter {
             return 5;
         }
 
-        return card.getAttack() + card.getDefense() / cost;
+        return (card.getAttack() + card.getDefense()) / (double) cost;
     }
 }
